requirements.txt
absl-py==2.1.0
Brotli==1.0.9
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.1
cycler==0.11.0
einops==0.8.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.62.2
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
kiwisolver==1.4.8
Markdown==3.4.1
MarkupSafe==2.1.5
matplotlib==3.10.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.2.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==24.2
pillow==11.1.0
pip==25.0
portalocker==3.1.1
protobuf==4.25.3
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
safetensors==0.6.2
setuptools==72.1.0
sip==6.10.0
six==1.16.0
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard_data_server==0.7.0
termcolor==2.5.0
timm==1.0.19
tomli==2.0.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yacs==0.1.8
